+++ app/src/test/java/protect/card_locker/ImportExportTest.java
@@ -980,7 +980,9 @@ public class ImportExportTest
                 "1,Card 1,Note 1,1618053234,100,USD,1234,5432,1,QR_CODE,0,\r\n" +
                 "8,Clothes Store,Note about store,,0,,a,,-5317,,0,\n" +
                 "2,Department Store,,1618041729,0,,A,,-9977996,,0,\n" +
-                "3,Grocery Store,,,150,,dhd,,-9977996,,0,\n" +
+                "3,Grocery Store,\"Multiline note about grocery store\n" +
+                "\n" +
+                "with blank line\",,150,,dhd,,-9977996,,0,\n" +
                 "4,Pharmacy,,,0,,dhshsvshs,,-10902850,,1,\n" +
                 "5,Restaurant,Note about restaurant here,,0,,98765432,23456,-10902850,CODE_128,0,\n" +
                 "6,Shoe Store,,,12.50,EUR,a,-5317,,AZTEC,0,\n" +
@@ -1065,7 +1067,7 @@ public class ImportExportTest
         LoyaltyCard card3 = db.getLoyaltyCard(3);
 
         assertEquals("Grocery Store", card3.store);
-        assertEquals("", card3.note);
+        assertEquals("Multiline note about grocery store\n\nwith blank line", card3.note);
         assertEquals(null, card3.expiry);
         assertEquals(new BigDecimal("150"), card3.balance);
         assertEquals(null, card3.balanceType);